requirements.txt
allure-pytest==2.13.5
pytest==8.3.3
pytest-rerunfailures==14.0
requests==2.32.3
